requirements.txt
gradio
sentence-transformers
faiss-cpu
transformers
torch
numpy